Updated the files/merged
Merged multiple files together into partial classes. Hoping this will help ensure process flows for the current code is easy to follow when altering for VR. Further organizing can come later
