+++ CSS551_FinalProject_RayMichael/Assets/UI/MainController.cs
@@ -6,8 +6,8 @@ public partial class MainController : MonoBehaviour
 {
     public Camera mainCamera = null;
     public TheWorld mModel = null;
-    public TheWorld2 mModel2 = null;
-    public TheWorldController mModel3 = null;
+    //public TheWorld2 mModel2 = null;
+    //public TheWorldController mModel3 = null;
 
     private GameObject mSelected;
     private Vector3 clawPos = new Vector3(0.0f, 4f, 0.0f);
@@ -44,7 +44,7 @@ public partial class MainController : MonoBehaviour
     void Start()
     {
         Debug.Assert(mainCamera != null);
-        Debug.Assert(mModel2 != null);
+        //Debug.Assert(mModel2 != null);
         Debug.Assert(LookAt != null);
 
         Debug.Assert(handle != null);
@@ -67,18 +67,17 @@ public partial class MainController : MonoBehaviour
         {
             LiftClaw();
         }
+
+        ClawMovement();
     }
 
     private void LMB()
     {
-
         if (Input.GetMouseButtonDown(0))
         {
             mousPosX = Input.mousePosition.x;
             mousPosY = Input.mousePosition.y;
-            Debug.Log("Mouse Position is : " + mousPosX + ", " + mousPosY);
 
-            Debug.Log("Clicked the left mouse button");
             RaycastHit hitInfo = new RaycastHit();
             bool hit = Physics.Raycast(Camera.main.ScreenPointToRay(Input.mousePosition), out hitInfo, Mathf.Infinity);
             if (hit)
@@ -87,21 +86,19 @@ public partial class MainController : MonoBehaviour
                 Debug.Log(hitInfo.point);
                 if (ComputeHandleDetection(hitInfo.point))
                 {
-                    Debug.Log("You hit the handle!");
                     handleSelected = true;
                     prevMousePos = mainCamera.ScreenToWorldPoint(Input.mousePosition + depthBuffer);
                 }
                 else if (ComputeDropDetection(hitInfo.point))
                 {
-                    Debug.Log("You hit the drop button!");
                     Drop = true;
                     mModel.clawActionFlag = "drop";
-                    mModel3.PushButton(0);
+                    mModel.PushButton(0);
                 }
                 else if (ComputeResetDetection(hitInfo.point))
                 {
                     Debug.Log("You hit the reset button!");
-                    mModel3.PushButton(1);
+                    mModel.PushButton(1);
                     ResetClaw();
                 }
             }
@@ -137,7 +134,7 @@ public partial class MainController : MonoBehaviour
             Vector3 p2 = r.GetPoint(100.0f);
 
             Vector3 initPos = handle.GetLocalPosition();
-            mModel3.UpdateJointRotation(p2, initPos);
+            mModel.UpdateJointRotation(p2, initPos);
 
             //DragHandle();
         }
@@ -145,7 +142,7 @@ public partial class MainController : MonoBehaviour
 
     private void DragHandle() {         
         Vector3 currentMousePos = mainCamera.ScreenToWorldPoint(Input.mousePosition + depthBuffer);
-        mModel3.UpdateJointRotation2(prevMousePos, currentMousePos);
+        mModel.UpdateJointRotation2(prevMousePos, currentMousePos);
     }
 
     private bool ComputeHandleDetection(Vector3 pos2)
@@ -163,8 +160,7 @@ public partial class MainController : MonoBehaviour
         Vector3 scaleHandle = handle.GetLocalScale();
         float r = scaleHandle.x * 0.5f;
 
-        Vector3 ph;
-        float d, a;
+        float d;
 
         d = Mathf.Sqrt(X.sqrMagnitude - (h * h));
         if (d < r)
@@ -190,8 +186,7 @@ public partial class MainController : MonoBehaviour
         Vector3 scaleHandle = dropBtn.GetLocalScale();
         float r = scaleHandle.x * 0.5f;
 
-        Vector3 ph;
-        float d, a;
+        float d;
 
         d = Mathf.Sqrt(X.sqrMagnitude - (h * h));
         if (d < r)
@@ -217,130 +212,13 @@ public partial class MainController : MonoBehaviour
         Vector3 scaleHandle = resetBtn.GetLocalScale();
         float r = scaleHandle.x * 0.5f;
 
-        Vector3 ph;
-        float d, a;
+        float d;
 
         d = Mathf.Sqrt(X.sqrMagnitude - (h * h));
         if (d < r)
         {
             hit = true;
         }
-
         return hit;
-    }
-
-
-    private void CraneMovement()
-    {
-        if (Input.GetKey(KeyCode.UpArrow))
-        {
-            //Move the crane claw "up" in positive Z direction
-            if (clawPos.z < backWall)
-            {
-                Vector3 pos = new Vector3();
-                pos.z = pos.z + speed * 1.0f;
-                clawPos += pos;
-                // Debug.Log(clawPos);
-                //clawPos = mModel.UpdateClawPosition(clawPos);
-            }
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-
-        }
-        else if (Input.GetKey(KeyCode.LeftArrow))
-        {
-            //Move the crane claw "left" in negative X direction
-            if (clawPos.x > leftWall)
-            {
-                Vector3 pos = new Vector3();
-                pos.x = pos.x + speed * -1.0f;
-                clawPos += pos;
-                // Debug.Log(clawPos);
-                //clawPos = mModel.UpdateClawPosition(clawPos);
-            }
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-        }
-        else if (Input.GetKey(KeyCode.RightArrow))
-        {
-            //Move the crane claw "right" in positive X direction
-            if (clawPos.x < rightWall)
-            {
-                Vector3 pos = new Vector3();
-                pos.x = pos.x + speed * 1.0f;
-                clawPos += pos;
-                // Debug.Log(clawPos);
-                //clawPos = mModel.UpdateClawPosition(clawPos);
-            }
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-        }
-        else if (Input.GetKey(KeyCode.DownArrow))
-        {
-            //Move the crane claw "down" in negative Z direction
-            if (clawPos.z > frontWall)
-            {
-                Vector3 pos = new Vector3();
-                pos.z = pos.z + speed * -1.0f;
-                clawPos += pos;
-                // Debug.Log(clawPos);
-            }
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-        }
-    }
-
-    private void DropClaw()
-    {
-        //Drop the crane claw in negative Y direction
-        timer += Time.deltaTime;
-        // Debug.Log(timer);
-
-        
-        //Vector3 pos = new Vector3();
-       //pos.y = clawPos.y;
-        if (clawPos.y > 1.5f)
-        {
-            Vector3 pos = new Vector3();
-            pos.y = pos.y + (speed + 0.75f) * -1.0f * Time.deltaTime;
-            clawPos.y += pos.y;
-            // Debug.Log(clawPos);
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-            Debug.Log("Dropping the claw");
-        }
-        else if (mModel.clawActionFlag == "drop") {
-            mModel.clawActionFlag = "close";
-        }
-        else if (mModel.clawActionFlag == "raise")
-        {
-            mModel2.GrabPrize();
-            Drop = false;
-            Lift = true;
-        }        
-    }
-
-    private void LiftClaw()
-    {
-        if ((clawPos.y > 0.0f) && (clawPos.y < 4f))
-        {
-            Vector3 pos = new Vector3();
-            pos.y = pos.y + (speed + 0.75f) * 1.0f * Time.deltaTime;
-            clawPos.y += pos.y;
-            // Debug.Log(clawPos);
-            clawPos = mModel2.UpdateClawPosition(clawPos);
-            Debug.Log("Lifting the claw");
-        }
-        else
-        {
-            Lift = false;
-        }
-    }
-
-    private void ResetClaw()
-    {
-        //Reset the claw to original position
-        Vector3 orig = new Vector3();
-        orig.y = 4.0f;
-        orig.x = 0.0f;
-        orig.z = 0.0f;
-        clawPos = mModel2.UpdateClawPosition(orig);
-        
-        Drop = false;
     }
 }
